User methods transferred to independent Controller, sqlController file deleted
